refactor: Comment 도메인 디렉토리 변경
- Article에 포함관계를 고려해 변경
