+++ src/main/java/com/sedin/qna/article/comment/model/Comment.java
@@ -1,4 +1,4 @@
-package com.sedin.qna.comment.model;
+package com.sedin.qna.article.comment.model;
 
 import com.sedin.qna.account.model.Account;
 import com.sedin.qna.article.model.Article;
